reusable config reading
